Check if file exists

+++ api/images/rocky-9/entrypoint.sh
@@ -1,7 +1,9 @@
 #!/bin/sh
 set -e
 
-cp /etc/halon/tls/ca-crt/tls.crt /etc/pki/ca-trust/source/anchors/halon-ca.crt
-update-ca-trust extract
+if [ -f /etc/halon/tls/ca-crt/tls.crt ]; then
+    cp /etc/halon/tls/ca-crt/tls.crt /etc/pki/ca-trust/source/anchors/halon-ca.crt
+    update-ca-trust extract
+fi
 
 exec "$@"
