+++ clusterd/images/ubuntu-24.04/entrypoint.sh
@@ -1,7 +1,9 @@
 #!/bin/sh
 set -e
 
-cp /etc/halon/tls/ca-crt/tls.crt /usr/local/share/ca-certificates/halon-ca.crt
-update-ca-certificates
+if [ -f /etc/halon/tls/ca-crt/tls.crt ]; then
+    cp /etc/halon/tls/ca-crt/tls.crt /usr/local/share/ca-certificates/halon-ca.crt
+    update-ca-certificates
+fi
 
 exec "$@"
